Spliting the parsing functions out, so they can be used as a library.
